added city abbreviation
so literally you just have to run this script and it should make the city_attributes table. aka the one with all the lat/long values and city abbreviations. The abbr column could be used as a means of seperating the user interface layer from the business logic, if I were able to then go forth and change all the table columns to just be the abbreviations.