adjusted donut chart, scraping work
Just added to the donut chart; colorborders so that we can create more seperation between the sections, but it doesn't seem to work.. Otherwise this commit has all of my work scraping from today. city_pics_url.json is just a dump of my mongoDB while city_scrape.ipynb is all the work done to build it.